pyproject.toml
[tool.poetry]
name = "sentinel-cognitive-forge"
version = "5.0.0"
description = "State-of-the-art Cognitive Forge system with advanced AI capabilities."
authors = ["Sentinel AI Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
crewai = "^0.36.1"
crewai-tools = "^0.3.0"
langchain-google-genai = "^1.0.6"
chromadb = "^0.5.0"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
sqlalchemy = "^2.0.30"
pysqlite3-binary = "^0.5.2"
psutil = "^5.9.8"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.45.0"}
numpy = "^1.26.4"
pydantic = "^2.6.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
asyncio-mqtt = "^0.16.0"
watchfiles = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
